requirements.txt
streamlit
pdfplumber
python-docx
gspread
google-auth
openai
python-dotenv
